# NAIS_PRD/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NAIS AI-OS / Research Commons API
  version: 1.9.0
  description: |
    P0 contract. Owner: Agent 0 (Platform). Changes only via CONTRACT_CHANGE_REQUEST.md.

    Conventions
    - Base path `/api/v1`, served behind gateway `http://localhost:21051`.
    - IDs are opaque UUID (v7 recommended). Times are UTC ISO 8601.
    - Errors use `ErrorEnvelope`; codes are listed in `contracts/error_codes.json`.
    - List endpoints use cursor pagination (`cursor`, `limit` default 20, max 100).
    - Every response carries `X-Request-Id`; `error.trace_id` equals the OpenTelemetry trace id.
    - "Not visible" resources return 404 (not 403) to avoid existence leaks, except where noted.
servers:
  - url: http://localhost:21051/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [projects]
      operationId: listProjects
      description: Projects the caller is a member of, plus PUBLIC projects when `scope=discover`.
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, discover], default: mine } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ProjectStatus" } }
        - { name: q, in: query, schema: { type: string } }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageEnvelope"
                  - { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/ProjectSummary" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    post:
      tags: [projects]
      operationId: createProject
      description: Caller becomes PROJECT_OWNER; caller's organization becomes lead organization.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ProjectCreate" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /projects/{project_id}:
    parameters: [ { $ref: "#/components/parameters/ProjectId" } ]
    get:
      tags: [projects]
      operationId: getProject
      description: Members only. Non-members get 404 for PRIVATE and 403 FORBIDDEN for PUBLIC.
      responses:
        "200": { description: Project, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [projects]
      operationId: updateProject
      description: PROJECT_OWNER / PROJECT_ADMIN.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ProjectUpdate" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /projects/{project_id}/archive:
    parameters: [ { $ref: "#/components/parameters/ProjectId" } ]
    post:
      tags: [projects]
      operationId: archiveProject
      description: PROJECT_OWNER only. Emits project.archived.v1; governance revokes all project-scoped grants.
      responses:
        "200": { description: Archived, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /projects/{project_id}/members:
    parameters: [ { $ref: "#/components/parameters/ProjectId" } ]
    get:
      tags: [projects]
      operationId: listProjectMembers
      responses:
        "200": { description: Members, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/ProjectMember" } } } } } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [projects]
      operationId: addProjectMember
      description: PROJECT_OWNER / PROJECT_ADMIN. Cross-organization users allowed. P0 adds directly (ACTIVE) and notifies (D-007).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_id, role]
              properties:
                user_id: { $ref: "#/components/schemas/Id" }
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "201": { description: Added, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectMember" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /projects/{project_id}/members/{user_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/UserId"
    patch:
      tags: [projects]
      operationId: updateProjectMemberRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectMember" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [projects]
      operationId: removeProjectMember
      description: PROJECT_OWNER / PROJECT_ADMIN, or the member themself (leave). Emits project.member.removed.v1.
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }

  # ---------------------------------------------------------------- catalog (M03)
...
  /internal/demo/projects:
    put:
      tags: [Internal]
      operationId: syncDemoProjects
      security: [ { internalToken: [] } ]
      description: |
        Makes the demo project index (alias nais-demo-projects, never the real nais-projects) equal to the given public project summaries; the api adds
        the embedding. Token rule as getInternalNotebookActivity. Index not configured or unreachable -> 503 DEPENDENCY_UNAVAILABLE.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InternalDemoProjectsRequest" }
      responses:
        "200":
          description: What was written
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InternalDemoSyncResult" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /internal/demo/projects/search:
    post:
      tags: [Internal]
      operationId: searchDemoProjects
      security: [ { internalToken: [] } ]
      description: |
        Ids of the demo projects that match q by words (BM25) or by meaning (k-NN), best first. Token rule as getInternalNotebookActivity.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InternalDemoProjectSearchRequest" }
      responses:
        "200":
          description: Ranked project ids
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InternalDemoProjectSearchResult" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak realm `nais` access token. Required claims sub, email, name, org_code; audience nais-api.
    internalToken:
      type: apiKey
      in: header
      name: X-NAIS-Internal-Token
      description: Shared server-to-server secret NAIS_INTERNAL_TOKEN (D-049); Internal operations only.
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    UserId: { name: user_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
    ProjectId: { name: project_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
...
  responses:
    Error:
      description: Error envelope. See contracts/error_codes.json for codes and HTTP mapping.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    Id: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code: { type: string, examples: ["ACCESS_GRANT_EXPIRED"] }
            message: { type: string }
            trace_id: { type: string }
            details: { type: object, additionalProperties: true }
    PageEnvelope:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: {} }
        page:
          type: object
          required: [has_more]
          properties:
            next_cursor: { type: [string, "null"] }
            has_more: { type: boolean }
...
    ProjectRole: { type: string, enum: [PROJECT_OWNER, PROJECT_ADMIN, RESEARCHER, VIEWER] }
    ProjectStatus: { type: string, enum: [ACTIVE, ARCHIVED] }
    ProjectVisibility: { type: string, enum: [PRIVATE, PUBLIC] }
...
    ProjectSummary:
      type: object
      required: [project_id, name, status, lead_organization_id, visibility]
      properties:
        project_id: { $ref: "#/components/schemas/Id" }
        name: { type: string }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        visibility: { $ref: "#/components/schemas/ProjectVisibility" }
        lead_organization_id: { $ref: "#/components/schemas/Id" }
        my_role: { oneOf: [ { $ref: "#/components/schemas/ProjectRole" }, { type: "null" } ] }
        lead_organization_name: { type: string }
        member_count: { type: integer }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
    Project:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          required: [description, organizations, created_at]
          properties:
            description: { type: string }
            keywords: { type: array, items: { type: string } }
            start_date: { type: [string, "null"], format: date }
            end_date: { type: [string, "null"], format: date }
            organizations:
              type: array
              items:
                type: object
                required: [organization_id, role]
                properties:
                  organization_id: { $ref: "#/components/schemas/Id" }
                  name: { type: string }
                  role: { type: string, enum: [LEAD, PARTNER], description: "PARTNER rows are derived from members' organizations" }
            created_by: { $ref: "#/components/schemas/Id" }
            created_at: { $ref: "#/components/schemas/Timestamp" }
            archived_at: { type: [string, "null"], format: date-time }
    ProjectCreate:
      type: object
      additionalProperties: false
      required: [name, description]
      properties:
        name: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, maxLength: 10000 }
        visibility: { $ref: "#/components/schemas/ProjectVisibility" }
        keywords: { type: array, items: { type: string, maxLength: 50 }, maxItems: 20 }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
    ProjectUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, maxLength: 10000 }
        visibility: { $ref: "#/components/schemas/ProjectVisibility" }
        keywords: { type: array, items: { type: string, maxLength: 50 }, maxItems: 20 }
        start_date: { type: [string, "null"], format: date }
        end_date: { type: [string, "null"], format: date }
    ProjectMember:
      type: object
      required: [project_id, user_id, organization_id, role, joined_at]
      properties:
        project_id: { $ref: "#/components/schemas/Id" }
        user_id: { $ref: "#/components/schemas/Id" }
        display_name: { type: string }
        organization_id: { $ref: "#/components/schemas/Id" }
        organization_name: { type: string }
        role: { $ref: "#/components/schemas/ProjectRole" }
        joined_at: { $ref: "#/components/schemas/Timestamp" }
        added_by: { $ref: "#/components/schemas/Id" }

    # ---------- catalog
...
    InternalDemoSyncResult:
      type: object
      required: [indexed, removed]
      properties:
        indexed: { type: integer, minimum: 0 }
        removed: { type: integer, minimum: 0 }
...
    InternalDemoProjectsRequest:
      type: object
      required: [documents]
      properties:
        documents:
          type: array
          maxItems: 2000
          items:
            type: object
            required: [project_id, name, lead_organization_id, updated_at]
            properties:
              project_id: { $ref: "#/components/schemas/Id" }
              name: { type: string, minLength: 1, maxLength: 200 }
              lead_organization_id: { $ref: "#/components/schemas/Id" }
              lead_organization_name: { type: string }
              member_count: { type: integer, minimum: 0 }
              updated_at: { $ref: "#/components/schemas/Timestamp" }
    InternalDemoProjectSearchRequest:
      type: object
      required: [q]
      properties:
        q: { type: string, minLength: 1, maxLength: 200 }
        limit: { type: integer, minimum: 1, maximum: 100, default: 20 }
    InternalDemoProjectSearchResult:
      type: object
      required: [project_ids]
      properties:
        project_ids: { type: array, items: { $ref: "#/components/schemas/Id" } }
...